docs(brew): prettify, remove existing deps, add warnings
macOS is Darwin based, which is POSIX-compatible via a BSD compatibility layer, this provides it with a lot of pre-existing utils such as tar, curl, zip, unzip, etc. Installing them from Homebrew is moot, and also discouraged by Homebrew itself.

Additionally, the guide uses a lot of fixed paths that the users need to be warned about, due to Homebrew's differences in installation between the two versionf of macOS (x86 and Apple Silicon).

Signed-off-by: rishyak <hello@rishyak.com>